fix(e2e): harden candidate evidence selection
